clean up the code
